// influx_ui/src/lib/types/AnnotatedDocument.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DocumentConstituent } from "./DocumentConstituent";
import type { Phrase } from "./Phrase";
import type { Token } from "./Token";

export interface AnnotatedDocument {
  type: "AnnotatedDocument";
  text: string;
  constituents: Array<DocumentConstituent>;
  num_sentences: number;
  num_tokens: number;
  orthography_set: Array<string>;
  lemma_set: Array<string>;
  token_dict: Record<string, Token> | null;
  phrase_dict: Record<string, Phrase> | null;
}